+++ tests/unit/Mvc/Model/CriteriaTest.php
@@ -51,11 +51,12 @@ class CriteriaTest extends UnitTest
     /**
      * Tests Criteria::inWhere with empty array.
      *
+     * @test
      * @issue  10676
      * @author Serghei Iakovlev <serghei@phalconphp.com>
      * @since  2016-08-11
      */
-    public function testShouldExecuteInWhereQueryWithEmptyArray()
+    public function shouldExecuteInWhereQueryWithEmptyArray()
     {
         $this->specify(
             'The Criteria::inWhere with empty array does not work as expected',
@@ -66,5 +67,43 @@ class CriteriaTest extends UnitTest
                 expect($criteria->execute())->isInstanceOf(Simple::class);
             }
         );
+    }
+
+    /**
+     * Tests work with limit / offset
+     *
+     * @test
+     * @issue  12419
+     * @author Serghei Iakovelv <serghei@phalconphp.com>
+     * @since  2016-12-18
+     */
+    public function shouldCorrectHandleLimitAndOffset()
+    {
+        $this->specify(
+            'The criteria object works with limit / offset incorrectly',
+            function ($limit, $offset, $expected) {
+                /** @var \Phalcon\Mvc\Model\Criteria $query */
+                $query = Users::query();
+
+                $query->limit($limit, $offset);
+
+                expect($query->getLimit())->equals($expected);
+            },
+            ['examples' => $this->limitOffsetProvider()]
+        );
+    }
+
+    protected function limitOffsetProvider()
+    {
+        return [
+            [-7,      null,  7                                  ],
+            ["-7234", null,  7234                               ],
+            ["18",    null,  18                                 ],
+            ["18",    2,     ['number' => 18, 'offset' => 2]    ],
+            ["-1000", -200,  ['number' => 1000, 'offset' => 200]],
+            ["1000", "-200", ['number' => 1000, 'offset' => 200]],
+            ["0",    "-200", null                               ],
+            ["%3CMETA%20HTTP-EQUIV%3D%22refresh%22%20CONT ENT%3D%220%3Burl%3Djavascript%3Aqss%3D7%22%3E", 50, null],
+        ];
     }
 }